4.21*: bloglist expansion, step9

+++ part4/blog/controllers/blogs.js
@@ -47,15 +47,13 @@ blogsRouter.delete('/:id', async (request, response, next) => {
     }
     const user = await User.findById(decodedToken.id)
     const blog = await Blog.findById(request.params.id)
-    // console.log(user.id.toString())
-    // console.log(blog.user.toString())
     try {
-        // if (user.id.toString() === blog.user.toString()) {
-        await Blog.findByIdAndRemove(request.params.id)
-        response.status(204).end()
-        // } else {
-        //     return response.status(400).json({ error: 'can\'t delete posts you are not the owner of' })
-        // }
+        if (user.id.toString() === blog.user.toString()) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            return response.status(401).json({ error: 'Unauthorized can\'t delete posts you are not the owner of' })
+        }
     } catch (error) {
         next(error)
     }
